api/generate/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-firestore
google-auth
